backend/requirements.txt
# Web framework
fastapi==0.110.1
uvicorn[standard]==0.29.0

# AWS SDKs
boto3==1.34.71
#aioboto3==12.2.0  # Async-compatible AWS SDK (no aiobotocore needed)

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation and typing
pydantic==2.6.4
typing-extensions==4.10.0

# Optional database
sqlalchemy==2.0.30
databases==0.9.0
python-dotenv


# Async support and uploads
websockets==12.0
aiofiles==23.2.1
python-multipart==0.0.9

# Optional: Explainable AI
shap==0.45.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.8.4
